refactor: extract trailer and favorite buttons from VideoDetails
